Module Cellular_automata.Two

module type BSRules = sig ... end

BSRules contains information about the gamerules for a 2D CA

module B3_S23 : BSRules

B3_S23 contains the gamerules for Conway's Game of Life

module B36_S23 : BSRules

B36_S23 contains the gamerules for Highlife

module B3678_S34678 : BSRules

B3678_S34678 contains the gamerules for Day and Night

module B2_S : BSRules

B2_S contains the gamerules for Seeds

module type Board = sig ... end

Board contains the neccessary functions to observe 2D CA acorrding to specific gamerules

module MakeBoard (_ : BSRules) : Board

MakeBoard makes a Board according to the rules specified by BSRules